Pirates and Merchants
A gaming simulation similar to
“Warcraft,” consisting of objects located in a two-dimensional world
that move around and behave in various ways. The user enters commands
to tell the objects what to do, and the
Objects behave in simulated time. Simulated time advances one “tick” or
unit at a time. Time is “frozen” while the user enters commands. When
the user commands the program to “go”, time will advance one tick of
time. When the user commands the program to “run”, time will advance
several units of time (5 to be exact) until some significant event
happens such as docking at a port, picking up luggage etc.